fix(ci): python-coverage-shards — split oversized pr_monitor test file under 1500-line guardrail
test_pr_monitor_part_001.py grew to 2195 lines, tripping
test_first_party_code_files_stay_under_line_limit (MAX_FIRST_PARTY_FILE_LINES=1500).
Move the TestCiFailure class into a new sibling test_pr_monitor_part_003.py,
mirroring the existing per-part helper duplication pattern, so each module
stays under the limit. No test behavior changed; all 146 cases still run.
